adds extended modal to upload achievement record removes achievement option selection from the course page
