+++ frontend-nx/apps/edu-hub/components/CourseContent/AchievementRecord/UploadAchievementRecordModal/AchievementOptionDropDown.tsx
@@ -3,9 +3,9 @@ import Fade from '@material-ui/core/Fade';
 import MaterialMenu, { MenuProps } from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import { withStyles } from '@material-ui/core/styles';
-import { MinAchievementOption } from '../../helpers/achievement';
+import { MinAchievementOption } from '../../../../helpers/achievement';
 import React, { FC, createElement, useCallback } from 'react';
-import { AchievementOptionCourses_AchievementOptionCourse_AchievementOption } from '../../queries/__generated__/AchievementOptionCourses';
+import { AchievementOptionCourses_AchievementOptionCourse_AchievementOption } from '../../../../queries/__generated__/AchievementOptionCourses';
 
 interface IProps {
   anchorElement: HTMLElement | undefined;
