+++ frontend-nx/apps/edu-hub/components/RegisterButton.tsx
@@ -1,18 +1,19 @@
-import { useRouter } from "next/router";
-import { FC, useCallback } from "react";
+import { useRouter } from 'next/router';
+import { FC, useCallback } from 'react';
 import useTranslation from 'next-translate/useTranslation';
 
-import { Button } from "./common/Button";
+import { Button } from './common/Button';
 
 export const RegisterButton: FC = () => {
-  const { t } = useTranslation("common");
+  const { t } = useTranslation('common');
   const router = useRouter();
 
   const register = useCallback(() => {
-    const url = `${process.env.NEXT_PUBLIC_AUTH_URL
-      }/realms/edu-hub/protocol/openid-connect/registrations?client_id=hasura&redirect_uri=${encodeURIComponent(
-        window.location.href
-      )}&response_type=code`;
+    const url = `${
+      process.env.NEXT_PUBLIC_AUTH_URL
+    }/realms/edu-hub/protocol/openid-connect/registrations?client_id=hasura&redirect_uri=${encodeURIComponent(
+      window.location.href
+    )}&response_type=code`;
 
     if (!url) return;
     router.push(new URL(url));
@@ -20,8 +21,7 @@ export const RegisterButton: FC = () => {
 
   return (
     <Button onClick={register} filled inverted>
-      {t("registerButton")}
+      {t('registerButton')}
     </Button>
   );
 };
-
