+++ frontend-nx/apps/edu-hub/locales/en/course-page.json
@@ -124,14 +124,15 @@
   "achievement-option": "Achievement Record",
   "achievement-record-upload-dead-line-text": "You have time until {{date}} to upload an achievement record.",
   "choose-achievement-option": "Choose the title of your achievement record.",
-  "current_achievement_option": "Current achievement record",
+  "selected_achievement_option": "Selected achievement record",
   "no_current_achievement_option": "Until now no achievement record was submitted.",
   "no-achievement-option": "Currently is no achievement record option available for this course.",
-  "upload-proof": "Upload proof",
-  "download-template": "Download Template",
-  "last-upload-from-an-author": "The last proof with you as the author was uploaded by {{fullName}} on {{dateTime}}.",
-  "no-proof-uploaded-by-me-as-author": "So far, no proof with you as the author has been uploaded.",
-  "confirm-online-course-description": "Hereby you confirm that you have participated in the required online courses. Your proof of performance is your reflection sheet. You can find it <component>HERE</component>. If you have any questions, please contact your course leader.",
+  "upload_achievement_record": "Upload new record",
+  "use_documentation_template": "Use the following documentation template",
+  "download_template": "Download Template",
+  "last_record_upload": "The last record with you as the author was uploaded by {{fullName}} on {{dateTime}} for \"{{achievementRecordTitle}}\".",
+  "no_existing_record_upload": "So far, no record with you as the author has been uploaded.",
+  "confirm-online-course-description": "Hereby you confirm that you have participated in the required online courses. Your record of performance is your reflection sheet. You can find it <component>HERE</component>. If you have any questions, please contact your course leader.",
   "confirm-online-course": "Confirm",
   "online-course-confirmed": "You have confirmed on {{dateTime}} that you have participated in the required online courses.",
   "achievement-certificate-generation": "Generate Achievement Certificates",
@@ -141,5 +142,6 @@
   "no-certificates-generated": "No certificates have been generated.",
   "participationDataGenerate": "Generate Participation Data (CSV)",
   "participationDataDownload": "Download Participation Data (CSV)",
-  "selectTemplateDefaultLabel": "Choose template"
+  "selectTemplateDefaultLabel": "Choose template",
+  "achievement_option_not_selected": "Please select an achievement record option."
 }
