+++ frontend-nx/apps/edu-hub/pages/manage/users/index.tsx
@@ -21,9 +21,7 @@ const Users: FC = () => {
           <link rel="icon" href="/favicon.png" />
         </Head>
         <Page>
-          <div className="min-h-[77vh]">
-            {isLoggedIn && isAdmin && <UserDashboard t={t} />}
-          </div>
+          <div className="min-h-[77vh]">{isLoggedIn && isAdmin && <UserDashboard t={t} />}</div>
         </Page>
       </div>
     </>
@@ -36,9 +34,7 @@ interface IProps {
   t: any;
 }
 const UserDashboard: FC<IProps> = ({ t }) => {
-  const menuItems: StaticComponentProperty[] = [
-    { key: -1, label: t('all'), selected: true },
-  ];
+  const menuItems: StaticComponentProperty[] = [{ key: -1, label: t('all'), selected: true }];
   const [searchText, setSearchText] = useState('');
 
   /* #region Callbacks */
@@ -77,13 +73,7 @@ interface IMenubarProps {
   topMenuItems: StaticComponentProperty[];
   onMenuItemClick?: (property: StaticComponentProperty) => void;
 }
-const Menubar: FC<IMenubarProps> = ({
-  t,
-  topMenuItems,
-  handleSearch,
-  searchText,
-  onMenuItemClick,
-}) => {
+const Menubar: FC<IMenubarProps> = ({ t, topMenuItems, handleSearch, searchText, onMenuItemClick }) => {
   const [menuItems, setMenuItems] = useState(topMenuItems);
 
   const updateMenuBar = useCallback(
@@ -111,18 +101,10 @@ const Menubar: FC<IMenubarProps> = ({
     <div className="flex justify-between mb-5 text-white">
       <div className="flex items-center space-x-5">
         {menuItems.map((tab) => (
-          <EhMenuItem
-            key={tab.key}
-            property={tab}
-            onClickCallback={handleTabClick}
-          />
+          <EhMenuItem key={tab.key} property={tab} onClickCallback={handleTabClick} />
         ))}
       </div>
-      <SearchBox
-        onChangeCallback={handleSearch}
-        placeholder={t('userSearchPlaceHolder')}
-        searchText={searchText}
-      />
+      <SearchBox onChangeCallback={handleSearch} placeholder={t('userSearchPlaceHolder')} searchText={searchText} />
     </div>
   );
 };
